# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mare Nostrum API",
        "description": "Smart Mobility Flow — Batam-Singapore cross-border logistics tracking API.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get paginated list of companies",
                "operationId": "8af1177352a157e9228c00f023616d87",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "internal",
                                "partner"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of companies retrieved successfully."
                    }
                }
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Create a new company",
                "operationId": "dd2ae66c40ea62cd8f20d49bc08e4b21",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "city",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Company B Logistics"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "partner",
                                        "enum": [
                                            "internal",
                                            "partner"
                                        ]
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Batam"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Kawasan Industri Mukakuning"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.08234
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 104.03214
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Company created successfully."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                }
            }
        },
        "/companies/{id}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get single company details",
                "operationId": "662432d4be675d549b84ca1bf7e56b8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company details retrieved successfully."
                    },
                    "404": {
                        "description": "Company not found."
                    }
                }
            },
            "put": {
                "tags": [
                    "Companies"
                ],
                "summary": "Update existing company details",
                "operationId": "00dbc7963c0f6ac45bd2b2595fb1f88e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Company B Logistics Updated"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "internal",
                                            "partner"
                                        ]
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company updated successfully."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Companies"
                ],
                "summary": "Delete a company",
                "operationId": "00a9e816ead34effffa2a6378e1cde4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company deleted successfully."
                    },
                    "422": {
                        "description": "Cannot delete company referenced in trips."
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Send as: Authorization: Bearer {token}",
                "bearerFormat": "Sanctum personal access token",
                "scheme": "bearer"
            }
        }
    },
...
}